requirements.txt
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.9.0
sounddevice>=0.4.0
pyaudio>=0.2.11
python-rtmidi>=1.4.0
pydmx>=1.0.0
matplotlib>=3.5.0
colorama>=0.4.4
pyyaml>=6.0
threading
time
json
math

